rename vars
