+++ ui/src/pages/namespace/Explorer/Page/poc/PageCompiler/primitives/keyValue/utils.ts
@@ -1,0 +1,11 @@
+import { KeyValueType } from "../../../schema/primitives/keyValue";
+import { useStringInterpolation } from "../Variable/utils/useStringInterpolation";
+
+export const useKeyValueArrayResolver = () => {
+  const interpolateString = useStringInterpolation();
+  return (input: KeyValueType[]): KeyValueType[] =>
+    input.map(({ key, value }) => {
+      const parsedValue = interpolateString(value);
+      return { key, value: parsedValue };
+    });
+};
