+++ ui/src/pages/namespace/Explorer/Page/poc/PageCompiler/procedures/utils.ts
@@ -1,31 +1,25 @@
-import { KeyValueType } from "../../schema/primitives/keyValue";
 import { MutationType } from "../../schema/procedures/mutation";
 import { QueryType } from "../../schema/procedures/query";
+import { useKeyValueArrayResolver } from "../primitives/keyValue/utils";
 import { useStringInterpolation } from "../primitives/Variable/utils/useStringInterpolation";
 
-export const useGetUrl = () => {
+export const useUrlGenerator = () => {
   const resolveKeyValueArray = useKeyValueArrayResolver();
   const interpolateString = useStringInterpolation();
 
   return (input: QueryType | MutationType) => {
     const { baseUrl, queryParams } = input;
-    const queryParamsParsed = resolveKeyValueArray(queryParams ?? []);
+    const queryParamsResolved = resolveKeyValueArray(queryParams ?? []);
     const searchParams = new URLSearchParams();
-    queryParamsParsed?.forEach(({ key, value }) => {
+    queryParamsResolved?.forEach(({ key, value }) => {
       searchParams.append(key, value);
     });
 
     const queryString = searchParams.toString();
-    const url = interpolateString(baseUrl);
-    return queryString ? url.concat("?", queryString) : url;
-  };
-};
+    const interpolatedBaseUrl = interpolateString(baseUrl);
 
-const useKeyValueArrayResolver = () => {
-  const interpolateString = useStringInterpolation();
-  return (input: KeyValueType[]): KeyValueType[] =>
-    input.map(({ key, value }) => {
-      const parsedValue = interpolateString(value);
-      return { key, value: parsedValue };
-    });
+    return queryString
+      ? interpolatedBaseUrl.concat("?", queryString)
+      : interpolatedBaseUrl;
+  };
 };
